queue-worker: move scrape worker to separate threads
